Fix migration

+++ lib/classes/patch/390alpha7a.php
@@ -16,6 +16,7 @@ use Alchemy\Phrasea\Model\Entities\FeedEntry;
 use Alchemy\Phrasea\Model\Entities\FeedItem;
 use Alchemy\Phrasea\Model\Entities\FeedPublisher;
 use Alchemy\Phrasea\Model\Entities\FeedToken;
+use Doctrine\ORM\Query\ResultSetMapping;
 
 class patch_390alpha7a implements patchInterface
 {
@@ -52,10 +53,33 @@ class patch_390alpha7a implements patchInterface
     /**
      * {@inheritdoc}
      */
+    public function getDoctrineMigrations()
+    {
+        return array('feed', 'aggregate-token');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function apply(base $appbox, Application $app)
     {
-        $version = $app['doctrine-migration.configuration']->getVersion($this->release);
-        $version->execute('up');
+        $rsm = new ResultSetMapping();
+        $rsm->addScalarResult('Name', 'Name');
+
+        $backup = false;
+
+        foreach ($app['EM']->createNativeQuery('SHOW TABLE STATUS', $rsm)->getResult() as $row) {
+            if (0 === strcmp('feeds', $row['Name'])) {
+                $backup = true;
+                break;
+            }
+        }
+
+        if (false === $backup) {
+            return false;
+        }
+
+        $app['EM']->executeQuery("RENAME TABLE `feeds` TO `feeds_backup`");
 
         $sql = 'DELETE FROM Feeds';
         $stmt = $app['phraseanet.appbox']->get_connection()->prepare($sql);
@@ -88,22 +112,6 @@ class patch_390alpha7a implements patchInterface
         $stmt->closeCursor();
 
         $conn = $app['phraseanet.appbox']->get_connection();
-
-        $sql = 'SHOW TABLE STATUS;';
-        $stmt = $conn->prepare($sql);
-        $stmt->execute();
-        $rs = $stmt->fetchAll(\PDO::FETCH_ASSOC);
-
-        $found = false;
-        foreach ($rs as $row) {
-            if ('feeds_backup' === $row['Name']) {
-                $found = true;
-                break;
-            }
-        }
-        if (!$found) {
-            return;
-        }
 
         $sql = 'SELECT id, title, subtitle, public, created_on, updated_on, base_id FROM feeds_backup;';
         $stmt = $conn->prepare($sql);
